+++ tests/test_engine/test_queries/test_queryop_comparsion_nin.py
@@ -3,7 +3,7 @@ import pytest
 import re
 
 from montydb.errors import OperationFailure
-from montydb.types import bson_ as bson
+from montydb.types import bson
 
 from ...conftest import skip_if_no_bson
