+++ tests/test_engine/test_queries/test_queryop_element_type.py
@@ -1,6 +1,6 @@
 
 import re
-from montydb.types import bson_ as bson
+from montydb.types import bson
 from datetime import datetime
 
 from ...conftest import skip_if_no_bson
