+++ tests/test_engine/test_update/test_update_max.py
@@ -1,5 +1,5 @@
 
-from montydb.types import bson_ as bson
+from montydb.types import bson
 from datetime import datetime
 
 from ...conftest import skip_if_no_bson
